Move UI utilities to cmd/ since they are only used there
